+++ api/twitch/twitch.controller.js
@@ -1,7 +1,17 @@
+// #region Imports
 /* Class imports */
 const crypto = require("crypto");
 const Logger = require("../../lib/logger");
 const NotifyDiscord = require("../../lib/notifyDiscord");
+
+/* Database Query import */
+const service = require("./twitch.service");
+
+/* Middleware import */
+const verifyInternalToken = require("../common/verifyAuthToken");
+const announcementExists = require("../common/announcementExists");
+const verifyAnnouncementResult = require("../common/verifyAnnouncementResult");
+const asyncErrorBoundary = require("../errors/asyncErrorBoundary");
 
 /* Get defined log level and instantiate a new logger */
 const { LOG_LEVEL } = process.env;
@@ -13,7 +23,9 @@ const { TWITCH_SECRET } = process.env;
 
 /* Initialize empty set for storing notification IDs */
 const TwitchNotifIDs = new Set();
+// #endregion Imports
 
+// #region Twitch EventSub Communication
 /**
  * After signature is verified,
  * checks the request message type header
@@ -82,14 +94,15 @@ function requestRouter(request, response, next) {
       const isUnique = isUniqueNotification(notificationID);
 
       /* Send "is duplicate" error if necessary */
-      if (isUnique === false)
+      if (isUnique === false) {
         next({ status: 400, message: "Duplicate payload detected!" });
+      } else {
+        /* Respond with a "green light" to Twitch (200 OK) */
+        response.sendStatus(200);
+      }
 
       /* Get the "meat" of the information from the notification payload */
       const { event } = request.body;
-
-      /* Respond with a "green light" to Twitch (200 OK) */
-      response.sendStatus(200);
 
       logger.info({
         action: "Notification recieved and validated",
@@ -180,7 +193,335 @@ function isUniqueNotification(notificationID) {
   /* Return boolean value held in isUnique */
   return isUnique;
 }
+// #endregion Twitch EventSub Communication
+
+// #region Set Announcement Channel for a Discord Server
+/**
+ * Sends either an INSERT or UPDATE request
+ * to the DB, depending on if an existing entry
+ * for the guild ID was located in the DB
+ * previously.
+ *
+ * @returns Status code and JSON object
+ */
+async function setAnnouncement(request, response) {
+  const { announcement } = response.locals;
+  const newChannelID = response.locals.channel_id;
+
+  logger.info({
+    action: "Set Announcement Channel",
+    location: `'setAnnouncement' in ${__dirname}`,
+    notes: [
+      `Announcement Object: ${announcement}`,
+      `Channel ID: ${newChannelID}`,
+    ],
+  });
+
+  /**
+   * INSERT into DB new announcement entry
+   * if existing entry was not found
+   * in previous middleware
+   */
+  if (announcement === null) {
+    const newGuildID = request.params.guild_id;
+    const newAnnouncement = {
+      guild_id: newGuildID,
+      channel_id: newChannelID,
+    };
+    /* INSERT */
+    const insertedAnnouncement = await service.createAnnouncement(
+      newAnnouncement
+    );
+
+    response.status(201).json({ data: await insertedAnnouncement });
+  } else {
+    /* UPDATE existing DB by guild_id with new channel_id */
+    const guild_id = announcement.guild_id;
+    const updatedAnnouncement = await service.updateAnnouncement(
+      guild_id,
+      newChannelID
+    );
+
+    response.status(200).json({ data: await updatedAnnouncement });
+  }
+}
+
+/**
+ * Checks request query parameters for the new
+ * channel ID
+ */
+function hasChannelID(request, response, next) {
+  logger.trace({
+    action: "Verify Channel ID is Present",
+    location: `'hasChannelID' in ${__dirname}`,
+  });
+
+  /* Get channel ID from query parameters */
+  const { channel_id } = request.query;
+
+  /* Error message snippet for formatting purposes */
+  const errorQuery = "?channel_id=<channel ID>";
+
+  /* If no channel ID, return an error */
+  if (!channel_id)
+    next({
+      status: 400,
+      message: `Channel ID missing! Please enter it as a query parameter as follows: ${errorQuery}`,
+    });
+
+  /* Continue */
+  logger.info({
+    action: "Channel ID is Present",
+    location: `'hasChannelID' in ${__dirname}`,
+  });
+
+  /* Store new channel ID in memory and continue */
+  response.locals.channel_id = channel_id;
+  next();
+}
+
+/**
+ * Verify that the user-provided channel ID is new,
+ * and not the same as the previously saved ID
+ * Will "quick continue" if entry was not found
+ * in DB
+ */
+function channelIDisNew(request, response, next) {
+  logger.trace({
+    action: "Verify Channel ID is New",
+    location: `'channelIDisNew' in ${__dirname}`,
+    notes: ["Will 'quick continue' if entry not found in DB"],
+  });
+
+  /* Quick continue if announcement not found in DB */
+  if (response.locals.announcement === null) {
+    logger.info({
+      action: "No Entry found in DB - Continue to Next Middleware Early",
+      location: `'channelIDisNew' in ${__dirname}`,
+    });
+    next();
+  } else {
+    const oldID = response.locals.announcement.channel_id;
+    const newID = response.locals.channel_id;
+
+    if (oldID == newID)
+      next({
+        status: 202,
+        message: `Channel ID is the same as previous ID. No changes.`,
+      });
+
+    /* Channel ID confirmed new, continue */
+    logger.info({
+      action: "Channel ID Confirmed New",
+      location: `'channelIDisNew' in ${__dirname}`,
+      notes: [`New Channel ID: ${newID}`],
+    });
+    next();
+  }
+}
+// #endregion Set Announcement Channel for a Discord Server
+
+// #region Get Announcement Channel
+/**
+ * Mainly just used as validation that an announcement
+ * channel has been set by the user
+ * @returns {Object} announcement channel information
+ */
+function getAnnouncementChannel(request, response) {
+  logger.info({
+    action: "Return Announcement Object",
+    location: `"getAnnouncementChannel" in ${__dirname}`,
+  });
+
+  /* Get announcement object from memory */
+  const { announcement } = response.locals;
+
+  response.status(200).json({ data: announcement });
+}
+// #endregion Get Announcement Channel
+
+// #region Add New Twitch Broadcaster Subscription for a Discord Server
+/**
+ * Sends INSERT request to DB with new
+ * subscription
+ * @returns Status code and JSON object
+ */
+async function createNewLogEntry(request, response) {
+  /* Retrieve subscription information from memory */
+  const { broadcaster_id } = response.locals;
+  const { guild_id } = response.locals.announcement;
+
+  /* Set up new subscription object to be inserted into DB */
+  const subscription = {
+    guild_id,
+    broadcaster_id,
+  };
+
+  /* Send INSERT query */
+  const newLogEntry = await service.createSubscription(subscription);
+
+  /* Respond to user with result */
+  response.status(201).json({ data: await newLogEntry });
+}
+
+/**
+ * Checks request query parameters for
+ * broadcaster ID
+ */
+function hasBroadcasterID(request, response, next) {
+  logger.trace({
+    action: "Verify Broadcaster ID is Present",
+    location: `'hasBroadcasterID' in ${__dirname}`,
+  });
+
+  /* Get broadcaster ID from query parameters */
+  const { broadcaster_id } = request.query;
+
+  /* Error message snippet for formatting purposes */
+  const errorQuery = "?broadcaster_id=<broadcaster ID>";
+
+  /* If no broadcaster ID, return an error */
+  if (!broadcaster_id)
+    next({
+      status: 400,
+      message: `Broadcaster ID missing! Please enter it as a query parameter as follows: ${errorQuery}`,
+    });
+
+  /* Continue */
+  logger.info({
+    action: "Broadcaster ID is Present",
+    location: `'hasBroadcasterID' in ${__dirname}`,
+    notes: [`Broadcaster ID: ${broadcaster_id}`],
+  });
+
+  response.locals.broadcaster_id = broadcaster_id;
+  next();
+}
+
+/**
+ * Verifies that the subscription request
+ * will not create a duplicate entry in the
+ * database
+ */
+async function isNotDuplicateSub(request, response, next) {
+  logger.trace({
+    action: "Check for Duplicate Subscription",
+    location: `'isNotDuplicateSub' in ${__dirname}`,
+  });
+
+  const { guild_id } = request.params;
+  const { broadcaster_id } = response.locals;
+
+  const checkSubscription = await service.readSub(guild_id, broadcaster_id);
+
+  if (checkSubscription) {
+    next({
+      status: 409,
+      message: `There is already a subscription in that guild for broadcaster '${broadcaster_id}'`,
+    });
+  } else {
+    logger.info({
+      action: "No Duplicate Subscription Found",
+      location: `'isNotDuplicateSub' in ${__dirname}`,
+      notes: [`${JSON.stringify(checkSubscription, null, 2)}`],
+    });
+    next();
+  }
+}
+// #endregion Add New Twitch Broadcaster Subscription for a Discord Server
+
+// #region Get All Channels for Given Broadcaster ID
+/**
+ * Retrieves all related channel IDs
+ * from the "subscriptions" table in DB
+ * for a given broadcaster ID
+ * @returns {Array} Channel IDs
+ */
+async function listChannelIDs(request, response) {
+  /* Get broadcaster ID */
+  const { broadcaster_id } = response.locals;
+
+  logger.trace({
+    action: "Get All Channel IDs for Given Broadcaster ID",
+    location: `'listChannelIDs' in ${__dirname}`,
+  });
+
+  /* Retrieve all related channel IDs */
+  const channelIDs = await service.listSubsForBroadcaster(broadcaster_id);
+
+  logger.info({
+    action: "Get All Channel IDs for Given Broadcaster ID",
+    location: `'listChannelIDs' in ${__dirname}`,
+    notes: [
+      `Broadcaster ID: ${broadcaster_id}`,
+      `Related Channel IDs: ${JSON.stringify(await channelIDs, null, 2)}`,
+    ],
+  });
+
+  /* Respond with array and status 200 OK */
+  response.status(200).json({ data: await channelIDs });
+}
+// #endregion Get All Channels for Given Broadcaster ID
+
+// #region Destroy a Subscription
+async function destroySub(request, response) {
+  const { guild_id } = request.params;
+  const { broadcaster_id } = response.locals;
+
+  /* Send delete request */
+  await service.unsubscribe(guild_id, broadcaster_id);
+
+  /* Respond with Success code 204 (No Content) */
+  response.sendStatus(204);
+}
+// #endregion Destroy a Subscription
+
+// #region List Active Subscriptions for a Guild
+async function listSubs(request, response) {
+  const { guild_id } = request.params;
+
+  const broadcasterIDs = await service.listSubsByGuild(guild_id);
+
+  logger.info({
+    action: "Get All Subscriptions for a Guild",
+    location: `'listSubs' in ${__dirname}`,
+    notes: [
+      `Broadcaster IDs: ${JSON.stringify(await broadcasterIDs, null, 2)}`,
+    ],
+  });
+
+  response.status(200).json({ data: await broadcasterIDs });
+}
+// #endregion List Active Subscriptions for a Guild
 
 module.exports = {
-  subscribe: [verifySignature, requestRouter],
+  eventsub: [verifySignature, requestRouter],
+  createSub: [
+    verifyInternalToken,
+    hasBroadcasterID,
+    asyncErrorBoundary(isNotDuplicateSub),
+    asyncErrorBoundary(announcementExists),
+    verifyAnnouncementResult,
+    asyncErrorBoundary(createNewLogEntry),
+  ],
+  setAnnounce: [
+    verifyInternalToken,
+    hasChannelID,
+    asyncErrorBoundary(announcementExists),
+    channelIDisNew,
+    asyncErrorBoundary(setAnnouncement),
+  ],
+  getAnnounce: [
+    verifyInternalToken,
+    asyncErrorBoundary(announcementExists),
+    verifyAnnouncementResult,
+    getAnnouncementChannel,
+  ],
+  getChannels: [verifyInternalToken, hasBroadcasterID, listChannelIDs],
+  destroySub: [
+    verifyInternalToken,
+    hasBroadcasterID,
+    asyncErrorBoundary(destroySub),
+  ],
+  listSubs: [verifyInternalToken, asyncErrorBoundary(listSubs)],
 };